feat: refactor ui to include ad-hoc token type information
