+++ src/frontend/src/mixins/token.js
@@ -7,6 +7,15 @@ export const token = {
       // return tokenStr.match(/.{1,4}/g).join('-'); // For "1234-1234-1234"
       return tokenStr.match(/.{1,3}/g).join(" ") // For "123 412 341 234"
     },
+    tokenTypeLabel(tokenType) {
+      const labels = {
+        energy: this.$tc("words.energy"),
+        time: this.$tc("words.time"),
+        unlock: this.$tc("words.unlock"),
+        reset: this.$tc("words.reset"),
+      }
+      return labels[tokenType] || tokenType
+    },
     unitLabel(unit) {
       const labels = {
         currency: this.$tc("words.currency"),
