+++ src/frontend/src/modules/Transactions/Transactions.vue
@@ -342,7 +342,7 @@
                     :md-label="$tc('words.type')"
                     md-sort-by="original_transaction_type"
                   >
-                    {{ formatTransactionType(item.type) }}
+                    {{ transactionTypeLabel(item.type) }}
                   </md-table-cell>
 
                   <md-table-cell
@@ -481,6 +481,7 @@ import waveMoneyLogo from "@/assets/icons/WaveMoney.png"
 import { currency } from "@/mixins/currency.js"
 import { notify } from "@/mixins/notify.js"
 import { timing } from "@/mixins/timing.js"
+import { transactionType } from "@/mixins/transactionType.js"
 import FilterTransaction from "@/modules/Transactions/FilterTransaction.vue"
 import { resources } from "@/resources.js"
 import { MainSettingsService } from "@/services/MainSettingsService.js"
@@ -493,7 +494,7 @@ import Widget from "@/shared/Widget.vue"
 
 export default {
   name: "Transactions.vue",
-  mixins: [timing, currency, notify],
+  mixins: [timing, currency, notify, transactionType],
   components: { Box, FilterTransaction, Widget },
   data() {
     return {
@@ -556,15 +557,6 @@ export default {
     EventBus.$off("searching", this.searching)
   },
   methods: {
-    formatTransactionType(type) {
-      const labels = {
-        energy: this.$tc("words.energy"),
-        deferred_payment: this.$tc("phrases.deferredPayment"),
-        eaas_rate: this.$tc("phrases.eaasRate"),
-        down_payment: this.$tc("phrases.downPayment"),
-      }
-      return labels[type] || type
-    },
     checkRouteChanges() {
       let isFiltering = false
       let queryParams = this.$route.query
